fix: Correctly position TransformationEditor on editor open and close states.
