+++ playwright.config.ts
@@ -23,13 +23,10 @@ export default defineConfig({
   reporter: 'html',
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
   use: {
-    /* Base URL to use in actions like `await page.goto('/')`. */
-    // baseURL: 'http://127.0.0.1:3000',
-
+    baseURL: process.env.PLAYWRIGHT_TEST_BASE_URL || 'http://localhost:5173/',
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
     trace: 'on-first-retry',
-    video: 'on',
-    baseURL: process.env.PLAYWRIGHT_TEST_BASE_URL || 'http://localhost:5173/'
+    video: 'on'
   },
   /* Configure projects for major browsers */
   projects: [
